// frontend/web/angular/src/app/generated/model/specialist.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Appointment } from './appointment';
import { Planing } from './planing';
import { SocialMediaLinks } from './socialMediaLinks';
import { SpecialistSpeciality } from './specialistSpeciality';
import { User } from './user';

export interface Specialist { 
    appointmentsList?: Array<Appointment>;
    biography?: string;
    city?: string;
    createdOn?: Date;
    gender?: string;
    lastUpdateOn?: Date;
    planings?: Array<Planing>;
    socialMediaLinks?: string;
    socialMediaLinksObject?: SocialMediaLinks;
    specialistId?: number;
    specialistSpecialityList?: Array<SpecialistSpeciality>;
    specialitiesList?: Array<string>;
    status?: number;
    userId?: User;
    yearOfExperience?: number;
}